requirements.txt
torch==1.2.0
torchvision==0.4.0
numpy==1.18.2
Pillow==6.2.0
